chore: archive 28 broken pygame-era legacy files
These tests/ and tools/ files have genuine syntax errors on every Python
version (verified failing on 3.10 and 3.13). They are pygame-era migration
debris; the live game is Godot/GDScript under godot/. Confirmed none are
imported by live code (godot/, scripts/, remaining tools/) before moving.
Moved to archive/legacy-pygame/ preserving relative subpath.

Also fix the pre-commit exclude regex: it only matched 'archived/' and
'docs/archive/', so the real top-level archive/ (and legacy/) trees were
still being linted. Added archive/ and legacy/ so archived broken files
are no longer formatted/parsed by black/ruff/isort (and so the crude
check-pygame-dir hook stops false-positiving on 'legacy-pygame/' paths).
