--- tests/test_config_manager.py
@@ -1,567 +1,0 @@
-'''
-Tests for the game configuration system.
-
-Tests cover:
-- Default config generation and validation
-- Multiple config file management  
-- Config switching and persistence
-- Integration with existing game systems
-- Error handling and edge cases
-
-NOTE: Tests temporarily disabled due to config_manager module import issues.
-See GitHub issue: https://github.com/PipFoweraker/pdoom1/issues/config-manager-import-bug
-'''
-
-import unittest
-import tempfile
-import shutil
-import os
-import pytest
-from unittest.mock import patch
-
-# Import the config manager
-from src.services.config_manager import ConfigManager, get_current_config, initialize_config_system
-
-
-class TestConfigManager(unittest.TestCase):
-    '''Test cases for the ConfigManager class.'''
-    
-    def setUp(self):
-        '''Set up test environment with temporary directory.'''
-        self.test_dir = tempfile.mkdtemp()
-        self.config_manager = ConfigManager()
-        # Override the config directory to use our test directory
-        self.config_manager.CONFIG_DIR = os.path.join(self.test_dir, 'configs')
-        self.config_manager.CURRENT_CONFIG_FILE = os.path.join(self.test_dir, 'current_config.json')
-        self.config_manager._ensure_config_directory()
-    
-    def tearDown(self):
-        '''Clean up test environment.'''
-        shutil.rmtree(self.test_dir)
-    
-    def test_config_directory_creation(self):
-        '''Test that config directory is created when needed.'''
-        # Directory should be created in setUp
-        self.assertTrue(os.path.exists(self.config_manager.CONFIG_DIR))
-    
-    def test_default_config_generation(self):
-        '''Test that default config is generated with expected structure.'''
-        default_config = self.config_manager.get_default_config()
-        
-        # Check required top-level sections
-        required_sections = [
-            'config_name', 'description', 'version',
-            'starting_resources', 'action_points', 'resource_limits',
-            'milestones', 'ui', 'audio', 'tutorial', 'gameplay', 'advanced'
-        ]
-        
-        for section in required_sections:
-            self.assertIn(section, default_config)
-        
-        # Check specific game balance values
-        self.assertEqual(default_config['starting_resources']['money'], 100000)
-        self.assertEqual(default_config['starting_resources']['staff'], 2)
-        self.assertEqual(default_config['action_points']['base_ap_per_turn'], 3)
-        self.assertEqual(default_config['milestones']['manager_threshold'], 9)
-        
-        # Check boolean settings
-        self.assertTrue(default_config['tutorial']['tutorial_enabled'])
-        self.assertTrue(default_config['audio']['sound_enabled'])
-    
-    def test_create_default_config_if_needed(self):
-        '''Test default config file creation when none exists.'''
-        # Should create config and return True
-        created = self.config_manager.create_default_config_if_needed()
-        self.assertTrue(created)
-        
-        # Should exist now
-        default_path = self.config_manager._get_config_path('default')
-        self.assertTrue(os.path.exists(default_path))
-        
-        # Should not create again and return False
-        created_again = self.config_manager.create_default_config_if_needed()
-        self.assertFalse(created_again)
-    
-    def test_list_available_configs(self):
-        '''Test listing available configuration files.'''
-        # Initially should be empty
-        configs = self.config_manager.list_available_configs()
-        self.assertEqual(len(configs), 1)  # Default is created automatically
-        self.assertIn('default', configs)
-        
-        # Create additional configs
-        test_config = self.config_manager.get_default_config()
-        test_config['config_name'] = 'Test Config'
-        self.config_manager.save_config('test_config', test_config)
-        
-        configs = self.config_manager.list_available_configs()
-        self.assertEqual(len(configs), 2)
-        self.assertIn('default', configs)
-        self.assertIn('test_config', configs)
-    
-    def test_save_and_load_config(self):
-        '''Test saving and loading configuration files.'''
-        test_config = {
-            'config_name': 'Test Configuration',
-            'description': 'Test config for unit tests',
-            'starting_resources': {'money': 2000}
-        }
-        
-        # Save config
-        success = self.config_manager.save_config('test', test_config)
-        self.assertTrue(success)
-        
-        # Load config
-        loaded_config = self.config_manager.load_config('test')
-        self.assertIsNotNone(loaded_config)
-        self.assertEqual(loaded_config['config_name'], 'Test Configuration')
-        self.assertEqual(loaded_config['starting_resources']['money'], 2000)
-        
-        # Test loading non-existent config
-        missing_config = self.config_manager.load_config('nonexistent')
-        self.assertIsNone(missing_config)
-    
-    def test_config_switching(self):
-        '''Test switching between different configurations.'''
-        # Create test config
-        test_config = self.config_manager.get_default_config()
-        test_config['config_name'] = 'Test Config'
-        test_config['starting_resources']['money'] = 5000
-        self.config_manager.save_config('test', test_config)
-        
-        # Switch to test config
-        success = self.config_manager.switch_config('test')
-        self.assertTrue(success)
-        self.assertEqual(self.config_manager.get_current_config_name(), 'test')
-        
-        # Verify current config has test values
-        current = self.config_manager.get_current_config()
-        self.assertEqual(current['starting_resources']['money'], 5000)
-        
-        # Switch to non-existent config should fail
-        success = self.config_manager.switch_config('nonexistent')
-        self.assertFalse(success)
-        self.assertEqual(self.config_manager.get_current_config_name(), 'test')  # Should remain unchanged
-    
-    def test_current_config_persistence(self):
-        '''Test that current config selection persists across sessions.'''
-        # Create and switch to test config
-        test_config = self.config_manager.get_default_config()
-        self.config_manager.save_config('persistent_test', test_config)
-        self.config_manager.switch_config('persistent_test')
-        
-        # Create new config manager instance to simulate restart
-        new_config_manager = ConfigManager()
-        new_config_manager.CONFIG_DIR = self.config_manager.CONFIG_DIR
-        new_config_manager.CURRENT_CONFIG_FILE = self.config_manager.CURRENT_CONFIG_FILE
-        new_config_manager._load_current_config_selection()
-        
-        # Should remember the selected config
-        self.assertEqual(new_config_manager.get_current_config_name(), 'persistent_test')
-    
-    def test_create_config_copy(self):
-        '''Test creating copies of existing configurations.'''
-        # Create original config
-        original_config = self.config_manager.get_default_config()
-        original_config['config_name'] = 'Original Config'
-        self.config_manager.save_config('original', original_config)
-        
-        # Create copy
-        success = self.config_manager.create_config_copy('original', 'copy')
-        self.assertTrue(success)
-        
-        # Verify copy exists and has updated metadata
-        copy_config = self.config_manager.load_config('copy')
-        self.assertIsNotNone(copy_config)
-        self.assertIn('Copy of', copy_config['config_name'])
-        self.assertIn('Based on original', copy_config['description'])
-        
-        # Should preserve other settings
-        self.assertEqual(copy_config['starting_resources']['money'], 
-                        original_config['starting_resources']['money'])
-        
-        # Test copying non-existent config
-        success = self.config_manager.create_config_copy('nonexistent', 'copy2')
-        self.assertFalse(success)
-    
-    def test_delete_config(self):
-        '''Test deleting configuration files.'''
-        # Create test config
-        test_config = self.config_manager.get_default_config()
-        self.config_manager.save_config('deleteme', test_config)
-        
-        # Verify it exists
-        configs = self.config_manager.list_available_configs()
-        self.assertIn('deleteme', configs)
-        
-        # Delete it
-        success = self.config_manager.delete_config('deleteme')
-        self.assertTrue(success)
-        
-        # Verify it's gone
-        configs = self.config_manager.list_available_configs()
-        self.assertNotIn('deleteme', configs)
-        
-        # Test deleting non-existent config
-        success = self.config_manager.delete_config('nonexistent')
-        self.assertFalse(success)
-        
-        # Test that default config cannot be deleted
-        success = self.config_manager.delete_config('default')
-        self.assertFalse(success)
-    
-    def test_config_info_retrieval(self):
-        '''Test getting metadata about configurations.'''
-        # Create test config
-        test_config = self.config_manager.get_default_config()
-        test_config['config_name'] = 'Info Test Config'
-        test_config['description'] = 'Config for testing metadata'
-        test_config['version'] = '2.0.0'
-        self.config_manager.save_config('info_test', test_config)
-        
-        # Get config info
-        info = self.config_manager.get_config_info('info_test')
-        self.assertIsNotNone(info)
-        self.assertEqual(info['name'], 'Info Test Config')
-        self.assertEqual(info['description'], 'Config for testing metadata')
-        self.assertEqual(info['version'], '2.0.0')
-        
-        # Test getting info for non-existent config
-        info = self.config_manager.get_config_info('nonexistent')
-        self.assertIsNone(info)
-    
-    def test_error_handling(self):
-        '''Test error handling for file operations.'''
-        # Test saving to invalid path
-        with patch('builtins.open', side_effect=PermissionError('Access denied')):
-            success = self.config_manager.save_config('test', {})
-            self.assertFalse(success)
-        
-        # Test loading corrupted config file
-        corrupted_path = self.config_manager._get_config_path('corrupted')
-        with open(corrupted_path, 'w') as f:
-            f.write('invalid json content {')
-        
-        loaded_config = self.config_manager.load_config('corrupted')
-        self.assertIsNone(loaded_config)
-
-
-@pytest.mark.skip(reason='Config system integration bugs - See issue #config-integration-bug')
-class TestConfigSystemIntegration(unittest.TestCase):
-    '''Test integration of config system with existing game components.'''
-    
-    def test_initialize_config_system(self):
-        '''Test the initialization function.'''
-        with tempfile.TemporaryDirectory() as temp_dir:
-            config_dir = os.path.join(temp_dir, 'configs')
-            with patch.object(ConfigManager, 'CONFIG_DIR', config_dir):
-                # Create a new manager instance for testing
-                with patch('src.services.config_manager.config_manager') as mock_global_manager:
-                    test_manager = ConfigManager()
-                    mock_global_manager.create_default_config_if_needed = test_manager.create_default_config_if_needed
-                    
-                    # First call should create default config
-                    created = initialize_config_system()
-                    self.assertTrue(created)
-                    
-                    # Second call should not create (already exists)
-                    created = initialize_config_system()
-                    self.assertFalse(created)
-    
-    def test_get_current_config_function(self):
-        '''Test the convenience function for getting current config.'''
-        # Mock the global config manager
-        with patch('src.services.config_manager.config_manager') as mock_manager:
-            mock_config = {'test': 'value'}
-            mock_manager.get_current_config.return_value = mock_config
-            
-            result = get_current_config()
-            self.assertEqual(result, mock_config)
-            mock_manager.get_current_config.assert_called_once()
-
-
-@pytest.mark.skip(reason='Config error handling bugs - See issue #config-error-handling-bug')
-class TestConfigErrorHandling(unittest.TestCase):
-    '''Test error handling and edge cases in the config system.'''
-    
-    def test_config_directory_creation_failure(self):
-        '''Test handling when config directory cannot be created.'''
-        with tempfile.TemporaryDirectory() as temp_dir:
-            # Create a file where we want the directory to be
-            fake_dir_path = os.path.join(temp_dir, 'configs')
-            with open(fake_dir_path, 'w') as f:
-                f.write('blocking file')
-            
-            # Try to create config manager with blocked directory
-            with patch.object(ConfigManager, 'CONFIG_DIR', fake_dir_path):
-                manager = ConfigManager()
-                # Should not crash, should use fallback
-                self.assertIsInstance(manager, ConfigManager)
-    
-    def test_config_save_permission_denied(self):
-        '''Test handling when config files cannot be saved due to permissions.'''
-        with tempfile.TemporaryDirectory() as temp_dir:
-            config_dir = os.path.join(temp_dir, 'configs')
-            os.makedirs(config_dir)
-            
-            with patch.object(ConfigManager, 'CONFIG_DIR', config_dir):
-                manager = ConfigManager()
-                
-                # Mock open to raise PermissionError
-                with patch('builtins.open', side_effect=PermissionError('Access denied')):
-                    result = manager.save_config('test', {'test': 'data'})
-                    self.assertFalse(result)
-    
-    def test_config_load_corrupted_file(self):
-        '''Test handling when config files are corrupted.'''
-        with tempfile.TemporaryDirectory() as temp_dir:
-            config_dir = os.path.join(temp_dir, 'configs')
-            os.makedirs(config_dir)
-            
-            # Create corrupted config file
-            corrupted_path = os.path.join(config_dir, 'corrupted.json')
-            with open(corrupted_path, 'w') as f:
-                f.write('{'incomplete': json}')
-            
-            with patch.object(ConfigManager, 'CONFIG_DIR', config_dir):
-                manager = ConfigManager()
-                result = manager.load_config('corrupted')
-                self.assertIsNone(result)
-    
-    def test_config_load_empty_file(self):
-        '''Test handling when config files are empty.'''
-        with tempfile.TemporaryDirectory() as temp_dir:
-            config_dir = os.path.join(temp_dir, 'configs')
-            os.makedirs(config_dir)
-            
-            # Create empty config file
-            empty_path = os.path.join(config_dir, 'empty.json')
-            with open(empty_path, 'w') as f:
-                pass  # Empty file
-            
-            with patch.object(ConfigManager, 'CONFIG_DIR', config_dir):
-                manager = ConfigManager()
-                result = manager.load_config('empty')
-                self.assertIsNone(result)
-    
-    def test_config_load_binary_file(self):
-        '''Test handling when config files contain binary data.'''
-        with tempfile.TemporaryDirectory() as temp_dir:
-            config_dir = os.path.join(temp_dir, 'configs')
-            os.makedirs(config_dir)
-            
-            # Create binary file
-            binary_path = os.path.join(config_dir, 'binary.json')
-            with open(binary_path, 'wb') as f:
-                f.write(b'\x00\x01\x02\x03\x04\x05')
-            
-            with patch.object(ConfigManager, 'CONFIG_DIR', config_dir):
-                manager = ConfigManager()
-                result = manager.load_config('binary')
-                self.assertIsNone(result)
-    
-    def test_graceful_fallback_when_all_fails(self):
-        '''Test that system gracefully falls back to in-memory defaults when everything fails.'''
-        with tempfile.TemporaryDirectory() as temp_dir:
-            config_dir = os.path.join(temp_dir, 'configs')
-            current_config_path = os.path.join(temp_dir, 'current_config.json')
-            
-            with patch.object(ConfigManager, 'CONFIG_DIR', config_dir):
-                with patch.object(ConfigManager, 'CURRENT_CONFIG_FILE', current_config_path):
-                    # Mock file operations to fail, but not the tempfile creation
-                    def mock_open_side_effect(*args, **kwargs):
-                        # Allow tempfile operations but block config file operations
-                        if 'pdoom_configs' in str(args[0]):
-                            return unittest.mock.mock_open()(*args, **kwargs)
-                        raise OSError('No filesystem access')
-                    
-                    with patch('builtins.open', side_effect=mock_open_side_effect):
-                        with patch('os.makedirs', side_effect=OSError('Cannot create directories')):
-                            manager = ConfigManager()
-                            # Should still be able to get a config (in-memory default)
-                            config = manager.get_current_config()
-                            self.assertIsInstance(config, dict)
-                            self.assertIn('starting_resources', config)
-    
-    def test_config_switching_to_nonexistent(self):
-        '''Test switching to a config that doesn't exist.'''
-        with tempfile.TemporaryDirectory() as temp_dir:
-            config_dir = os.path.join(temp_dir, 'configs')
-            os.makedirs(config_dir)
-            current_config_path = os.path.join(temp_dir, 'current_config.json')
-            
-            with patch.object(ConfigManager, 'CONFIG_DIR', config_dir):
-                with patch.object(ConfigManager, 'CURRENT_CONFIG_FILE', current_config_path):
-                    manager = ConfigManager()
-                    # Ensure we start with default
-                    self.assertEqual(manager.get_current_config_name(), 'default')
-                    
-                    # Should return False when switching to nonexistent config
-                    result = manager.switch_config('nonexistent_config')
-                    self.assertFalse(result)
-                    # Should maintain current config
-                    self.assertEqual(manager.get_current_config_name(), 'default')
-    
-    def test_current_config_file_corruption(self):
-        '''Test handling when current_config.json is corrupted.'''
-        with tempfile.TemporaryDirectory() as temp_dir:
-            # Create corrupted current_config.json
-            current_config_path = os.path.join(temp_dir, 'current_config.json')
-            with open(current_config_path, 'w') as f:
-                f.write('invalid json{')
-            
-            with patch.object(ConfigManager, 'CONFIG_DIR', os.path.join(temp_dir, 'configs')):
-                with patch.object(ConfigManager, 'CURRENT_CONFIG_FILE', current_config_path):
-                    # Should not crash, should fall back to default
-                    manager = ConfigManager()
-                    self.assertEqual(manager.get_current_config_name(), 'default')
-    
-    def test_config_system_resilience(self):
-        '''Test that the config system remains functional even with multiple failures.'''
-        with tempfile.TemporaryDirectory() as temp_dir:
-            config_dir = os.path.join(temp_dir, 'configs')
-            current_config_path = os.path.join(temp_dir, 'current_config.json')
-            
-            with patch.object(ConfigManager, 'CONFIG_DIR', config_dir):
-                with patch.object(ConfigManager, 'CURRENT_CONFIG_FILE', current_config_path):
-                    manager = ConfigManager()
-                    
-                    # Create some valid configs first
-                    manager.save_config('good1', {'name': 'Good Config 1'})
-                    manager.save_config('good2', {'name': 'Good Config 2'})
-                    
-                    # Create some corrupted configs
-                    os.makedirs(config_dir, exist_ok=True)
-                    with open(os.path.join(config_dir, 'bad1.json'), 'w') as f:
-                        f.write('invalid json')
-                    with open(os.path.join(config_dir, 'bad2.json'), 'w') as f:
-                        f.write('')
-                    
-                    # List configs should only return valid ones
-                    configs = manager.list_available_configs()
-                    # Should have at least default, good1, good2
-                    self.assertIn('default', configs)
-                    self.assertIn('good1', configs)
-                    self.assertIn('good2', configs)
-                    # bad1 and bad2 might be listed but won't load
-                    
-                    # Switching to good configs should work
-                    self.assertTrue(manager.switch_config('good1'))
-                    self.assertEqual(manager.get_current_config_name(), 'good1')
-                    
-                    # Switching to bad configs should fail gracefully
-                    self.assertFalse(manager.switch_config('bad1'))
-                    # Should maintain previous good config
-                    self.assertEqual(manager.get_current_config_name(), 'good1')
-
-
-@pytest.mark.skip(reason='Config game balance bugs - See issue #config-balance-bug')
-class TestConfigGameBalanceValidation(unittest.TestCase):
-    '''Test that default config contains sensible game balance values.'''
-    
-    def setUp(self):
-        '''Set up config manager for testing.'''
-        self.config_manager = ConfigManager()
-    
-    def test_starting_resources_reasonable(self):
-        '''Test that starting resources are reasonable.'''
-        config = self.config_manager.get_default_config()
-        starting = config['starting_resources']
-        
-        # Money should be positive and reasonable
-        self.assertGreater(starting['money'], 0)
-        self.assertLessEqual(starting['money'], 100000)  # Current balanced amount
-        
-        # Staff should be a small positive number
-        self.assertGreaterEqual(starting['staff'], 1)
-        self.assertLess(starting['staff'], 10)
-        
-        # Reputation should be reasonable
-        self.assertGreater(starting['reputation'], 0)
-        self.assertLess(starting['reputation'], 100)
-        
-        # Doom should be manageable starting point
-        self.assertGreater(starting['doom'], 0)
-        self.assertLess(starting['doom'], 50)
-    
-    def test_action_points_balanced(self):
-        '''Test that action point settings are balanced.'''
-        config = self.config_manager.get_default_config()
-        ap = config['action_points']
-        
-        # Base AP should be reasonable
-        self.assertGreaterEqual(ap['base_ap_per_turn'], 2)
-        self.assertLessEqual(ap['base_ap_per_turn'], 5)
-        
-        # Bonuses should be meaningful but not excessive
-        self.assertGreater(ap['staff_ap_bonus'], 0)
-        self.assertLess(ap['staff_ap_bonus'], 2)
-        
-        self.assertGreater(ap['admin_ap_bonus'], ap['staff_ap_bonus'])
-        self.assertLess(ap['admin_ap_bonus'], 5)
-        
-        # Max AP should prevent excessive accumulation
-        self.assertGreater(ap['max_ap_per_turn'], ap['base_ap_per_turn'])
-        self.assertLess(ap['max_ap_per_turn'], 20)
-    
-    def test_milestones_progression(self):
-        '''Test that milestone thresholds create good progression.'''
-        config = self.config_manager.get_default_config()
-        milestones = config['milestones']
-        
-        # Manager threshold should be reasonable
-        self.assertGreater(milestones['manager_threshold'], config['starting_resources']['staff'])
-        self.assertLess(milestones['manager_threshold'], 20)
-        
-        # Board spending threshold should be significant
-        self.assertGreater(milestones['board_spending_threshold'], config['starting_resources']['money'])
-        
-        # Event unlock turns should be reasonable
-        self.assertGreater(milestones['enhanced_events_turn'], 1)
-        self.assertLess(milestones['enhanced_events_turn'], 20)
-        
-        self.assertGreater(milestones['scrollable_log_turn'], 1)
-        self.assertLess(milestones['scrollable_log_turn'], 15)
-    
-    def test_resource_limits_sensible(self):
-        '''Test that resource limits prevent overflow/underflow.'''
-        config = self.config_manager.get_default_config()
-        limits = config['resource_limits']
-        
-        # All limits should be positive
-        for limit_name, limit_value in limits.items():
-            self.assertGreater(limit_value, 0, f'{limit_name} limit should be positive')
-        
-        # Limits should be higher than starting values
-        starting = config['starting_resources']
-        self.assertGreater(limits['max_money'], starting['money'])
-        self.assertGreater(limits['max_staff'], starting['staff'])
-        self.assertGreater(limits['max_reputation'], starting['reputation'])
-        self.assertGreater(limits['max_doom'], starting['doom'])
-    
-    def test_ui_settings_valid(self):
-        '''Test that UI settings have valid values.'''
-        config = self.config_manager.get_default_config()
-        ui = config['ui']
-        
-        # Scale values should be reasonable
-        self.assertGreater(ui['window_scale'], 0.1)
-        self.assertLessEqual(ui['window_scale'], 2.0)
-        
-        self.assertGreater(ui['font_scale'], 0.1)
-        self.assertLessEqual(ui['font_scale'], 3.0)
-        
-        self.assertGreater(ui['animation_speed'], 0.1)
-        self.assertLessEqual(ui['animation_speed'], 5.0)
-        
-        # Tooltip delay should be reasonable
-        self.assertGreaterEqual(ui['tooltip_delay'], 0)
-        self.assertLess(ui['tooltip_delay'], 5000)  # 5 seconds max
-        
-        # Boolean settings should be boolean
-        self.assertIsInstance(ui['show_balance_changes'], bool)
-        self.assertIsInstance(ui['show_keyboard_shortcuts'], bool)
-
-
-if __name__ == '__main__':
-    unittest.main()
