--- tests/test_new_player_experience.py
@@ -1,202 +1,0 @@
-# !/usr/bin/env python3
-'''
-Test script for the new player experience functionality.
-Validates that the new system works correctly.
-'''
-
-import sys
-import os
-import unittest
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-
-
-class TestNewPlayerExperience(unittest.TestCase):
-    '''Test cases for new player experience functionality.'''
-    
-    def test_menu_items_updated(self):
-        '''Test that menu items match current implementation.'''
-        import main
-        expected_menu_items = ['Launch Lab', 'Launch with Custom Seed', 'Settings', 'Player Guide', 'View Leaderboard', 'Exit']
-        actual_menu_items = main.menu_items
-        self.assertEqual(actual_menu_items, expected_menu_items, 
-                        f'Expected {expected_menu_items}, got {actual_menu_items}')
-    
-    def test_state_variables_exist(self):
-        '''Test that new player experience state variables are defined.'''
-        import main
-        
-        # Check that the variables are defined
-        self.assertTrue(hasattr(main, 'npe_tutorial_enabled'), 'npe_tutorial_enabled not found')
-        self.assertTrue(hasattr(main, 'npe_intro_enabled'), 'npe_intro_enabled not found')
-        self.assertTrue(hasattr(main, 'npe_selected_item'), 'npe_selected_item not found')
-        
-        # Check initial values
-        self.assertEqual(main.npe_tutorial_enabled, False, 'npe_tutorial_enabled should start as False')
-        self.assertEqual(main.npe_intro_enabled, False, 'npe_intro_enabled should start as False')
-        self.assertEqual(main.npe_selected_item, 0, 'npe_selected_item should start as 0')
-    
-    def test_handler_functions_exist(self):
-        '''Test that handler functions are defined.'''
-        import main
-        
-        self.assertTrue(hasattr(main, 'handle_new_player_experience_click'), 
-                       'handle_new_player_experience_click not found')
-        self.assertTrue(hasattr(main, 'handle_new_player_experience_hover'), 
-                       'handle_new_player_experience_hover not found')
-        self.assertTrue(hasattr(main, 'handle_new_player_experience_keyboard'), 
-                       'handle_new_player_experience_keyboard not found')
-    
-    def test_ui_drawing_function_exists(self):
-        '''Test that UI drawing function is defined.'''
-        try:
-            from ui import draw_new_player_experience
-            self.assertIsNotNone(draw_new_player_experience)
-        except ImportError as e:
-            self.fail(f'Could not import draw_new_player_experience: {e}')
-    
-    def test_state_manipulation(self):
-        '''Test that state variables can be manipulated.'''
-        import main
-        
-        # Save original values
-        original_tutorial = main.npe_tutorial_enabled
-        original_intro = main.npe_intro_enabled
-        
-        try:
-            # Simulate enabling tutorial
-            main.npe_tutorial_enabled = True
-            self.assertTrue(main.npe_tutorial_enabled, 'Could not enable tutorial')
-            
-            # Simulate enabling intro
-            main.npe_intro_enabled = True
-            self.assertTrue(main.npe_intro_enabled, 'Could not enable intro')
-            
-        finally:
-            # Reset to original state
-            main.npe_tutorial_enabled = original_tutorial
-            main.npe_intro_enabled = original_intro
-    
-    def test_intro_message_generation(self):
-        '''Test that intro message can be generated.'''
-        import main
-        from src.core.game_state import GameState
-        
-        # Save original state
-        original_intro = main.npe_intro_enabled
-        
-        try:
-            # Enable intro and create game state
-            main.npe_intro_enabled = True
-            test_game_state = GameState('test-seed')
-            
-            # Test intro message components
-            startup_money = test_game_state.money
-            self.assertIsInstance(startup_money, int, 'Startup money should be an integer')
-            self.assertGreater(startup_money, 0, 'Startup money should be positive')
-            
-            # Test expected intro format
-            expected_intro = f'Doom is coming. You convinced a funder to give you ${startup_money:,}. Your job is to save the world. Good luck!'
-            self.assertIsInstance(expected_intro, str, 'Intro message should be a string')
-            self.assertIn('Doom is coming', expected_intro, 'Intro should mention doom')
-            self.assertIn('save the world', expected_intro, 'Intro should mention saving the world')
-            
-        finally:
-            # Reset original state
-            main.npe_intro_enabled = original_intro
-
-
-def test_new_player_experience():
-    '''Legacy function for backward compatibility.'''
-    print('[TEST] Testing New Player Experience System')
-    print('=' * 50)
-    
-    # Test 1: Verify the new menu item exists
-    try:
-        import main
-        expected_menu_items = ['New Player Experience', 'Launch with Custom Seed', 'Settings', 'Player Guide', 'Exit']
-        actual_menu_items = main.menu_items
-        assert actual_menu_items == expected_menu_items, f'Expected {expected_menu_items}, got {actual_menu_items}'
-        print('[PASS] Test 1 PASSED: Menu items updated correctly')
-    except Exception as e:
-        print(f'[FAIL] Test 1 FAILED: {e}')
-        return False
-    
-    # Test 2: Verify new player experience state variables exist
-    try:
-        # Check that the variables are defined
-        assert hasattr(main, 'npe_tutorial_enabled'), 'npe_tutorial_enabled not found'
-        assert hasattr(main, 'npe_intro_enabled'), 'npe_intro_enabled not found'
-        assert hasattr(main, 'npe_selected_item'), 'npe_selected_item not found'
-        
-        # Check initial values
-        assert main.npe_tutorial_enabled == False, 'npe_tutorial_enabled should start as False'
-        assert main.npe_intro_enabled == False, 'npe_intro_enabled should start as False'
-        assert main.npe_selected_item == 0, 'npe_selected_item should start as 0'
-        print('[PASS] Test 2 PASSED: New player experience state variables correct')
-    except Exception as e:
-        print(f'[FAIL] Test 2 FAILED: {e}')
-        return False
-    
-    # Test 3: Verify handler functions exist
-    try:
-        assert hasattr(main, 'handle_new_player_experience_click'), 'handle_new_player_experience_click not found'
-        assert hasattr(main, 'handle_new_player_experience_hover'), 'handle_new_player_experience_hover not found' 
-        assert hasattr(main, 'handle_new_player_experience_keyboard'), 'handle_new_player_experience_keyboard not found'
-        print('[PASS] Test 3 PASSED: Handler functions defined')
-    except Exception as e:
-        print(f'[FAIL] Test 3 FAILED: {e}')
-        return False
-    
-    # Test 4: Verify UI function exists
-    try:
-        print('[PASS] Test 4 PASSED: UI drawing function defined')
-    except Exception as e:
-        print(f'[FAIL] Test 4 FAILED: {e}')
-        return False
-    
-    # Test 5: Test state manipulation
-    try:
-        # Simulate enabling tutorial
-        main.npe_tutorial_enabled = True
-        assert main.npe_tutorial_enabled == True, 'Could not enable tutorial'
-        
-        # Simulate enabling intro
-        main.npe_intro_enabled = True
-        assert main.npe_intro_enabled == True, 'Could not enable intro'
-        
-        # Reset for clean state
-        main.npe_tutorial_enabled = False
-        main.npe_intro_enabled = False
-        print('[PASS] Test 5 PASSED: State manipulation works')
-    except Exception as e:
-        print(f'[FAIL] Test 5 FAILED: {e}')
-        return False
-    
-    # Test 6: Test intro message integration
-    try:
-        from src.core.game_state import GameState
-        
-        # Enable intro and create game state
-        main.npe_intro_enabled = True
-        test_game_state = GameState('test-seed')
-        
-        # Simulate the intro message logic
-        startup_money = test_game_state.money
-        expected_intro = f'Doom is coming. You convinced a funder to give you ${startup_money:,}. Your job is to save the world. Good luck!'
-        print(f'[INFO] Intro message would be: '{expected_intro}'')
-        
-        # Reset
-        main.npe_intro_enabled = False
-        print('[PASS] Test 6 PASSED: Intro message generation works')
-    except Exception as e:
-        print(f'[FAIL] Test 6 FAILED: {e}')
-        return False
-    
-    print('=' * 50)
-    print('[CELEBRATION] ALL TESTS PASSED! New Player Experience system is working correctly.')
-    return True
-
-
-if __name__ == '__main__':
-    # Run unittest suite
-    unittest.main()
