--- tests/test_office_cat.py
@@ -1,335 +1,0 @@
-import unittest
-import sys
-import os
-
-# Add the parent directory to the path so we can import our modules
-sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-
-'''
-Test the office cat feature - enhanced dev engagement system
-'''
-import unittest
-from src.core.game_state import GameState
-
-
-class TestOfficeCat(unittest.TestCase):
-    def setUp(self):
-        """Set up test game state."""
-        self.game_state = GameState(seed=42)
-    
-    def test_initial_cat_state(self):
-        """Test that cat is not adopted initially."""
-        self.assertFalse(getattr(self.game_state, 'office_cat_adopted', False))
-        self.assertEqual(getattr(self.game_state, 'office_cat_turns_with_5_staff', 0), 0)
-        self.assertFalse(getattr(self.game_state, 'office_cat_adoption_offered', False))
-    
-    def test_staff_tracking(self):
-        """Test that consecutive turns with 5+ staff are tracked correctly."""
-        # Initially no staff tracking
-        self.assertEqual(getattr(self.game_state, 'office_cat_turns_with_5_staff', 0), 0)
-        
-        # Set staff to 5 and ensure enough money for maintenance
-        self.game_state.staff = 5
-        self.game_state.money = 10000
-        self.game_state.end_turn()
-        # Reset turn processing flag
-        self.game_state.turn_processing = False
-        self.assertEqual(self.game_state.office_cat_turns_with_5_staff, 1)
-        
-        # Another turn with 5+ staff
-        self.game_state.staff = 6
-        self.game_state.end_turn()
-        # Reset turn processing flag
-        self.game_state.turn_processing = False
-        self.assertEqual(self.game_state.office_cat_turns_with_5_staff, 2)
-        
-        # Drop below 5 staff - should reset counter
-        self.game_state.staff = 4
-        self.game_state.end_turn()
-        # Reset turn processing flag
-        self.game_state.turn_processing = False
-        self.assertEqual(self.game_state.office_cat_turns_with_5_staff, 0)
-    
-    def test_cat_adoption_trigger_conditions(self):
-        """Test that cat adoption event triggers under correct conditions."""
-        # Set up conditions: 5+ staff for 5+ turns, enough money
-        self.game_state.staff = 5
-        self.game_state.money = 10000  # Plenty of money for maintenance and adoption
-        
-        # Advance 5 turns with 5+ staff
-        for i in range(5):
-            self.game_state.end_turn()
-            self.game_state.turn_processing = False  # Reset turn processing flag
-        
-        self.assertEqual(self.game_state.office_cat_turns_with_5_staff, 5)
-        
-        # Check if cat adoption event would trigger
-        # We can't easily test the event trigger directly, but we can test the conditions
-        from src.core.events import EVENTS
-        cat_event = None
-        for event in EVENTS:
-            if event.get("name") == "Mysterious Office Visitor":
-                cat_event = event
-                break
-        
-        self.assertIsNotNone(cat_event)
-        self.assertTrue(cat_event["trigger"](self.game_state))
-    
-    def test_cat_adoption_with_insufficient_funds(self):
-        """Test cat adoption behavior when player has insufficient funds."""
-        # Set up conditions but with insufficient money
-        self.game_state.staff = 5
-        self.game_state.money = 50  # Less than minimum $89
-        
-        # Advance 5 turns with 5+ staff
-        for i in range(5):
-            self.game_state.end_turn()
-        
-        # Cat adoption should not trigger due to insufficient funds
-        from src.core.events import EVENTS
-        cat_event = None
-        for event in EVENTS:
-            if event.get("name") == "Mysterious Office Visitor":
-                cat_event = event
-                break
-        
-        self.assertIsNotNone(cat_event)
-        self.assertFalse(cat_event["trigger"](self.game_state))
-    
-    def test_cat_adoption_costs(self):
-        """Test that cat adoption deducts appropriate costs."""
-        # Manually trigger cat adoption
-        self.game_state.staff = 5
-        self.game_state.money = 200
-        self.game_state.office_cat_turns_with_5_staff = 5
-        
-        from src.core.events import trigger_office_cat_adoption
-        initial_money = self.game_state.money
-        
-        trigger_office_cat_adoption(self.game_state)
-        
-        # Should have adopted the cat and spent money
-        self.assertTrue(self.game_state.office_cat_adopted)
-        self.assertEqual(self.game_state.money, initial_money - 89)  # Basic package cost
-    
-    def test_cat_upkeep_costs(self):
-        """Test that cat upkeep costs are applied each turn."""
-        # Adopt the cat first
-        self.game_state.office_cat_adopted = True
-        self.game_state.staff = 0  # No staff to avoid maintenance costs
-        initial_money = self.game_state.money
-        
-        # End turn and check upkeep costs
-        self.game_state.end_turn()
-        
-        expected_cost = 14  # Weekly cat food cost
-        self.assertEqual(self.game_state.money, initial_money - expected_cost)
-        self.assertEqual(self.game_state.office_cat_total_food_cost, expected_cost)
-    
-    def test_cat_petting_interaction(self):
-        """Test that cat petting works correctly."""
-        # Adopt the cat
-        self.game_state.office_cat_adopted = True
-        self.game_state.office_cat_position = (400, 300)
-        
-        # Test successful petting (within range)
-        result = self.game_state.pet_office_cat((420, 320))  # Within 32 pixel range
-        self.assertTrue(result)
-        self.assertEqual(self.game_state.office_cat_total_pets, 1)
-        self.assertEqual(self.game_state.office_cat_love_emoji_timer, 60)
-        
-        # Test failed petting (out of range)
-        result = self.game_state.pet_office_cat((100, 100))  # Far away
-        self.assertFalse(result)
-        self.assertEqual(self.game_state.office_cat_total_pets, 1)  # Should not increase
-    
-    def test_cat_doom_stages(self):
-        """Test that cat doom stages progress correctly."""
-        self.game_state.office_cat_adopted = True
-        
-        # Stage 0: Happy cat (low doom)
-        self.game_state.doom = 10
-        self.assertEqual(self.game_state.get_cat_doom_stage(), 0)
-        
-        # Stage 1: Concerned cat
-        self.game_state.doom = 30
-        self.assertEqual(self.game_state.get_cat_doom_stage(), 1)
-        
-        # Stage 2: Alert cat
-        self.game_state.doom = 50
-        self.assertEqual(self.game_state.get_cat_doom_stage(), 2)
-        
-        # Stage 3: Ominous cat
-        self.game_state.doom = 70
-        self.assertEqual(self.game_state.get_cat_doom_stage(), 3)
-        
-        # Stage 4: Terrifying doom cat
-        self.game_state.doom = 90
-        self.assertEqual(self.game_state.get_cat_doom_stage(), 4)
-    
-    def test_cat_position_updates(self):
-        """Test that cat position updates work correctly."""
-        self.game_state.office_cat_adopted = True
-        
-        # Set initial position and test movement towards target
-        self.game_state.office_cat_position = (0, 0)
-        screen_w, screen_h = 1200, 800
-        
-        self.game_state.update_cat_position(screen_w, screen_h)
-        
-        # Cat should move towards bottom-right area
-        new_x, new_y = self.game_state.office_cat_position
-        self.assertGreater(new_x, 0)  # Should move right
-        self.assertGreater(new_y, 0)  # Should move down
-    
-    def test_cat_statistics_tracking(self):
-        """Test that cat statistics are tracked for end game."""
-        self.game_state.office_cat_adopted = True
-        self.game_state.office_cat_position = (400, 300)
-        
-        # Test food cost tracking
-        initial_food_cost = self.game_state.office_cat_total_food_cost
-        self.game_state.end_turn()
-        self.assertGreater(self.game_state.office_cat_total_food_cost, initial_food_cost)
-        
-        # Test petting tracking
-        self.game_state.pet_office_cat((420, 320))
-        self.assertEqual(self.game_state.office_cat_total_pets, 1)
-        
-        self.game_state.pet_office_cat((410, 310))
-        self.assertEqual(self.game_state.office_cat_total_pets, 2)
-    
-    def test_cat_morale_benefits(self):
-        """Test that cat provides morale benefits."""
-        self.game_state.office_cat_adopted = True
-        self.game_state.doom = 50
-        
-        # Multiple turns to test random morale boost
-        initial_doom = self.game_state.doom
-        doom_reductions = 0
-        
-        # Run multiple turns and count doom reductions
-        for _ in range(20):  # Run enough turns to likely see at least one morale boost
-            pre_turn_doom = self.game_state.doom
-            self.game_state.end_turn()
-            # Check if doom was reduced (accounting for normal doom increases)
-            # We need to check the messages to see if morale boost occurred
-            for message in self.game_state.messages:
-                if "morale boost" in message:
-                    doom_reductions += 1
-                    break
-        
-        # Should have gotten at least some morale boosts over 20 turns
-        # (30% chance per turn, so very likely to get at least one)
-        # Note: This is a probabilistic test that might rarely fail
-        self.assertGreaterEqual(doom_reductions, 0)  # At least possible
-
-
-if __name__ == '__main__':
-    unittest.main()
-        '''Set up test game state'''
-        self.gs = GameState('test-cat-seed')
-    
-    def test_office_cat_initialization(self):
-        '''Test that office cat system initializes correctly'''
-        # Cat should not be adopted initially
-        self.assertFalse(self.gs.office_cat_adopted)
-        self.assertEqual(self.gs.office_cat_total_pets, 0)
-        self.assertEqual(self.gs.office_cat_total_food_cost, 0)
-    
-    def test_cat_adoption_event(self):
-        '''Test that cat adoption event works correctly'''
-        # Manually trigger the adoption event
-        self.gs._trigger_stray_cat_adoption()
-        
-        # Verify cat is now adopted
-        self.assertTrue(self.gs.office_cat_adopted)
-        self.assertTrue(self.gs.office_cats_adopted)  # Legacy compatibility
-        
-        # Check that setup costs were applied
-        self.assertLess(self.gs.money, 100000)  # Flea treatment cost applied (started with $100k)
-        self.assertGreater(self.gs.reputation, 0)  # Reputation boost applied
-        
-        # Check messages were added
-        self.assertGreater(len(self.gs.messages), 0)
-        cat_messages = [msg for msg in self.gs.messages if 'cat' in msg.lower() or '[CAT]' in msg]
-        self.assertGreater(len(cat_messages), 0)
-    
-    def test_cat_petting_mechanics(self):
-        '''Test that cat petting works correctly'''
-        # Adopt the cat first
-        self.gs._trigger_stray_cat_adoption()
-        
-        # Clear messages from adoption
-        self.gs.messages = []
-        
-        # Test petting the cat (simulate mouse click near cat position)
-        cat_x, cat_y = self.gs.office_cat_position
-        result = self.gs.pet_office_cat((cat_x, cat_y))
-        
-        # Verify petting worked
-        self.assertTrue(result)
-        self.assertEqual(self.gs.office_cat_total_pets, 1)
-        self.assertEqual(self.gs.office_cat_last_petted, self.gs.turn)
-        self.assertGreater(self.gs.office_cat_love_emoji_timer, 0)
-    
-    def test_cat_upkeep_costs(self):
-        '''Test that cat upkeep costs are applied correctly'''
-        # Adopt the cat first
-        self.gs._trigger_stray_cat_adoption()
-        
-        # Note starting money
-        money_before = self.gs.money
-        
-        # Advance to next turn to trigger upkeep
-        self.gs.end_turn()
-        
-        # Verify upkeep costs were applied
-        self.assertLess(self.gs.money, money_before)
-        self.assertGreater(self.gs.office_cat_total_food_cost, 0)
-        
-        # Check upkeep message was added
-        upkeep_messages = [msg for msg in self.gs.messages if 'Cat upkeep' in msg]
-        self.assertGreater(len(upkeep_messages), 0)
-    
-    def test_cat_doom_stages(self):
-        '''Test that cat doom stages work correctly'''
-        # Adopt the cat first
-        self.gs._trigger_stray_cat_adoption()
-        
-        # Test different doom levels
-        self.gs.doom = 0
-        self.assertEqual(self.gs.get_cat_doom_stage(), 0)  # Happy cat
-        
-        self.gs.doom = 20
-        self.assertEqual(self.gs.get_cat_doom_stage(), 1)  # Concerned cat
-        
-        self.gs.doom = 50
-        self.assertEqual(self.gs.get_cat_doom_stage(), 2)  # Alert cat
-        
-        self.gs.doom = 70
-        self.assertEqual(self.gs.get_cat_doom_stage(), 3)  # Ominous cat
-        
-        self.gs.doom = 90
-        self.assertEqual(self.gs.get_cat_doom_stage(), 4)  # Doom cat
-    
-    def test_cat_position_updates(self):
-        '''Test that cat position updates work correctly'''
-        # Adopt the cat first
-        self.gs._trigger_stray_cat_adoption()
-        
-        # Update cat position with screen dimensions
-        screen_w, screen_h = 800, 600
-        self.gs.office_cat_position
-        
-        self.gs.update_cat_position(screen_w, screen_h)
-        
-        # Position should be updated (cat moves towards target)
-        final_pos = self.gs.office_cat_position
-        # Position should be valid coordinates
-        self.assertIsInstance(final_pos[0], int)
-        self.assertIsInstance(final_pos[1], int)
-
-
-if __name__ == '__main__':
-    unittest.main()
